RF | Move tests
